:dizzy: /join/:slug moves to routes/join.js
