+++ lib/db.js
@@ -1,4 +1,6 @@
 const config = require('../config');
+const utils = require('../lib/utils');
+
 const pg = require('pg');
 const pool = pg.Pool({
 	connectionString: config.db_dsn
@@ -391,6 +393,78 @@ const self = {
 				WHERE id = $2
 			`, [context.id, person.id]);
 		}
+	},
+
+	join_context: (person, context_id, invited_by) => {
+		return new Promise(async (resolve, reject) => {
+
+			try {
+
+				await self.query(`
+					UPDATE person
+					SET context_id = $1
+					WHERE id = $2
+				`, [context_id, person.id]);
+
+				let member = await self.get_member(person, context_id);
+				if (member) {
+					return resolve(member);
+				}
+
+				let leave_slug = utils.random(16);
+				let invite_slug = utils.random(16);
+				let query;
+
+				if (invited_by) {
+					query = await self.query(`
+						INSERT INTO member
+						(person_id, context_id, leave_slug, invite_slug, invited_by, created, updated)
+						VALUES ($1, $2, $3, $4, $5, CURRENT_TIMESTAMP, CURRENT_TIMESTAMP)
+						RETURNING *
+					`, [person.id, context_id, leave_slug, invite_slug, invited_by]);
+				} else {
+					query = await self.query(`
+						INSERT INTO member
+						(person_id, context_id, leave_slug, invite_slug, created, updated)
+						VALUES ($1, $2, $3, $4, CURRENT_TIMESTAMP, CURRENT_TIMESTAMP)
+						RETURNING *
+					`, [person.id, context_id, leave_slug, invite_slug]);
+				}
+
+				member = query.rows[0];
+				return resolve(member);
+
+			} catch (err) {
+				reject(err);
+			}
+
+		});
+	},
+
+	get_invite: (slug) => {
+		return new Promise(async (resolve, reject) => {
+			try {
+				let query = await self.query(`
+					SELECT *
+					FROM member
+					WHERE active = true
+					  AND invite_slug = $1
+				`, [slug]);
+
+				let invite;
+
+				if (query.rows.length == 0) {
+					resolve(false);
+				} else {
+					invite = query.rows[0];
+					invite.person = await self.get_person(invite.person_id);
+					invite.context = await self.get_context(invite.context_id);
+					resolve(invite);
+				}
+			} catch(err) {
+				reject(err);
+			}
+		});
 	},
 
 	get_member: (person, context_id, include_inactive) => {
